requirements.txt
Flask>=1.0
psycopg2-binary>=2.7
gunicorn>=19.9
Flask-WTF>=0.14.2
flask_bcrypt>=0.7.1
flask-login>=0.4.1